Pretty large re-working of the scrubber/editor system to incorporate it into the dagger-managed object-graph.
This also includes a re-work of the Moe main class into an abstract, with a managed lifecycle.  This allows for the injection of scrubbers (and repo factories) via a set-binding, and then allows the Google local flavor to inject additional metadata scrubbers and repository types.

An effect of this change is that nearly all google-specific code is siloed into a []/ sub-package, which can simply be omitted from the open-source fileset (making migration into third_party vastly easier).

Metadata scrubber config is extracted into the call-stack, rather than being injected (or, as it was, yanked out of a static reference to the context).  This lets various configurations be at play during a given moe run, as is the case for magic, so scrubbers are stateless (at least in terms of the parsed config) and can inspect the config for pieces they need as they are invoked.  Note, it is null-tolerant, to permit the many cases where there is no scrubber config.  However, a key design change is that some staged metadata scrubbers may choose to process in the absence of a scrubber config.  This lets company-specific scrubbers potentially be always on by default, if so desired, to ensure that certain metadata scrubbings happen regardless of per-project configuration.

	Change on 2016/03/07 by cgruber <cgruber@google.com>

-------------
Created by MOE: https://github.com/google/moe
MOE_MIGRATED_REVID=116573642